+++ Tests/Fixtures/StringableObject.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Tdn\PhpTypes\Tests\Fixtures;
+
+/**
+ * Class StringableObject
+ * @package Tdn\PhpTypes\Tests\Fixtures
+ */
+class StringableObject
+{
+    /**
+     * @return string
+     */
+    public function __toString()
+    {
+        return 'foo';
+    }
+}