+++ Type/String.php
@@ -4,229 +4,77 @@ namespace Tdn\PhpTypes\Type;
 
 use Doctrine\Common\Inflector\Inflector;
 use Stringy\Stringy;
+use Tdn\PhpTypes\Exception\InvalidTransformationException;
 
 /**
  * Class String
  * @package Tdn\PhpTypes\Type
  */
-class String extends Stringy
+class String extends Stringy implements TypeInterface
 {
     /**
      * Mainly here for type hinting purposes...
      *
-     * @param mixed $str
+     * @param string $str
      * @param string $encoding
      *
-     * @throws \InvalidArgumentException when $str is not a string.
-     *
-     * @return \Tdn\PhpTypes\Type\String
+     * @return static
      */
-    public static function create($str, $encoding = 'UTF-8')
+    public static function create($str = '', $encoding = 'UTF-8')
     {
-        $type = gettype($str);
-        if ($type !== 'string') {
-            throw new \InvalidArgumentException(
-                sprintf(
-                    'Expected string got %s instead.',
-                    $type
-                )
-            );
-        }
-
-        return new static($str, $encoding);
+        return parent::create($str, $encoding);
     }
 
     /**
      * Pluralizes the string.
      *
-     * @return self
+     * @return static
      */
     public function pluralize()
     {
-        return self::create($this->getInflector()->pluralize((string) $this->str), $this->encoding);
+        return static::create($this->getInflector()->pluralize((string) $this->str), $this->encoding);
     }
 
     /**
      * Singularizes the string.
      *
-     * @return self
+     * @return static
      */
     public function singularize()
     {
-        return self::create($this->getInflector()->singularize((string) $this->str), $this->encoding);
+        return static::create($this->getInflector()->singularize((string) $this->str), $this->encoding);
     }
 
     /**
-     * Returns substring from beginning until first instance of subsStr.
+     * Returns position of the first occurrence of subStr null if not present.
      *
-     * @param string $subStr
-     * @param bool $excluding
-     * @param bool $caseSensitive
-     *
-     * @return \Tdn\PhpTypes\Type\String
-     */
-    public function subStrUntil($subStr, $excluding = false, $caseSensitive = false)
-    {
-        $fromSubStr = $this->str[0];
-        return $this->subStrBetween($fromSubStr, $subStr, false, $excluding, $caseSensitive);
-    }
-
-    /**
-     * Returns substring from first instance of subStr to end of string.
-     * @param $subStr
-     * @param bool $excluding
-     * @param bool $caseSensitive
-     *
-     * @return \Tdn\PhpTypes\Type\String
-     */
-    public function subStrAfter($subStr, $excluding = false, $caseSensitive = false)
-    {
-        return $this->subStrBetween($subStr, null, $excluding, false, $caseSensitive);
-    }
-
-    /**
-     * Returns substring between fromSubStr to toSubStr. End of string if toSubStr is not set.
-     *
-     * @param string $fromSubStr
-     * @param string $toSubStr
-     * @param bool $excludeFromSubStr
-     * @param bool $excludeToSubStr
-     * @param bool $caseSensitive
-     * @return self
-     */
-    public function subStrBetween(
-        $fromSubStr,
-        $toSubStr = '',
-        $excludeFromSubStr = false,
-        $excludeToSubStr = false,
-        $caseSensitive = false
-    ) {
-        $fromIndex = 0;
-        $toIndex = mb_strlen($this->str);
-        $str = self::create($this->str);
-        if ($str->contains($fromSubStr)) {
-            $fromIndex = $this->strpos($fromSubStr, 0, $caseSensitive);
-            $fromIndex = ($excludeFromSubStr) ? $fromIndex + mb_strlen($fromSubStr, $this->encoding) : $fromIndex;
-
-            if ($fromIndex < 0) {
-                throw new \LogicException('To cannot be before from.');
-            }
-
-            if (!empty($toSubStr) && $str->contains($toSubStr)) {
-                $toIndex = $this->strpos($toSubStr, $fromIndex, $caseSensitive);
-                $toIndex = ($excludeToSubStr) ?
-                    $toIndex - $fromIndex :  ($toIndex - $fromIndex) + mb_strlen($toSubStr, $this->encoding);
-            }
-        }
-
-        return ($toSubStr) ? $str->substr($fromIndex, $toIndex) : $str->substr($fromIndex);
-    }
-
-    /**
-     * Returns position of the first occurance of subStr null if not present.
-     * @param $subStr
-     * @param int $start
-     * @param bool $caseSensitive
+     * @param string $subStr        Substring
+     * @param int|null $offset           Chars to offset from start
+     * @param bool $caseSensitive   Enable case sensitivity
      *
      * @return int
      */
-    public function strpos($subStr, $start = 0, $caseSensitive = false)
+    public function strpos($subStr, $offset = null, $caseSensitive = false)
     {
-        $res = ($caseSensitive) ?
-            mb_strpos($this->str, addslashes($subStr), $start, $this->encoding) :
-            mb_stripos($this->str, addslashes($subStr), $start, $this->encoding);
-
-        return $res;
+        return ($caseSensitive) ?
+            mb_strpos($this->str, $subStr, $offset, $this->encoding) :
+            mb_stripos($this->str, $subStr, $offset, $this->encoding);
     }
 
     /**
-     * Adds a predefined number of indentation indentation spaces to string.
-     * If newlines are found, it will add number of spaces before each newline.
+     * Returns position of the last occurrence of subStr null if not present.
      *
-     * @param int $numSpaces
-     * @param int $padType
-     * @param bool $perNewline
+     * @param string $subStr        Substring
+     * @param int|null $offset           Chars to offset from start
+     * @param bool $caseSensitive   Enable case sensitivity
      *
-     * @return self
-     */
-    public function addIndent($numSpaces = null, $padType = STR_PAD_LEFT, $perNewline = false)
-    {
-        $numSpaces = ($numSpaces) ? $numSpaces : 4;
-        $str = self::create($this->str);
-        $spaces = str_repeat(' ', $numSpaces);
-        if ($str->contains("\n") !== false && $perNewline) {
-            $lines = explode("\n", trim($this->str));
-            foreach ($lines as &$line) {
-                $line = $this->lineStrPad($line, $padType, $spaces);
-            }
-            return self::create(implode("\n", $lines));
-        }
-
-        return String::create($this->lineStrPad($str, $padType, $spaces));
-    }
-
-    /**
-     * Gets current indentation length. Skips newlines.
-     * Ensures that all lines passed are indented equally, otherwise fails.
-     * Returns the number of indentation.
-     *
-     * @param null $str
      * @return int
      */
-    public function getIndentSize($str = null)
+    public function strrpos($subStr, $offset = null, $caseSensitive = false)
     {
-        $str = ($str) ? $str : self::create($this->str);
-        $counters = [];
-        $position = 0;
-        $line     = 0;
-
-        foreach ($str as $letter) {
-            //No need to add support for tabs since we want to follow PSR2.
-            if ($letter == ' ') {
-                //If we're in the middle of a string, do not count the space.
-                if (isset($str[$position - 1]) && ($str[$position - 1] != ' ' && $str[$position - 1] != "\n")) {
-                    continue;
-                }
-
-                $counters[$line]++;
-            }
-
-            //Make sure not to initiate a new array
-            if ($letter == "\n" && isset($str[$position + 1])) {
-                $line++;
-                $counters[$line] = 0;
-            }
-
-            $position++;
-        }
-
-        if (count(array_unique($counters)) > 1) {
-            throw new \RuntimeException('String passed is not correctly indented. Indentation is not consistent.');
-        }
-
-        return array_pop($counters);
-    }
-
-    /**
-     * @param $line
-     * @param $padType
-     * @param $spaces
-     * @return string
-     */
-    private function lineStrPad($line, $padType, $spaces)
-    {
-        switch($padType) {
-            case STR_PAD_LEFT:
-                $line = $spaces . $line;
-                break;
-            case STR_PAD_BOTH:
-                $line = $spaces . $line . $spaces;
-                break;
-            case STR_PAD_RIGHT:
-                $line = $line . $spaces;
-        }
-
-        return $line;
+        return ($caseSensitive) ?
+            mb_strrpos($this->str, $subStr, $offset, $this->encoding) :
+            mb_strripos($this->str, $subStr, $offset, $this->encoding);
     }
 
     /**
@@ -243,5 +91,46 @@ class String extends Stringy
     public function __toString()
     {
         return (string) $this->str;
+    }
+
+    /**
+     * @param mixed $mixed
+     *
+     * @return static
+     */
+    public static function valueOf($mixed)
+    {
+        switch (strtolower(gettype($mixed))) {
+            case 'boolean':
+                return new static((($mixed) ? 'true' : 'false'));
+            case 'integer':
+            case 'float':
+            case 'double':
+            case 'string':
+                return new static((string) $mixed);
+            case 'array':
+                return new static(implode(', ', $mixed));
+            case 'object':
+                if (method_exists($mixed, '__toString')) {
+                    return new static((string) $mixed);
+                }
+
+                throw new InvalidTransformationException(
+                    sprintf(
+                        'Could not transform %s to string.',
+                        get_class($mixed)
+                    )
+                );
+            case 'resource':
+                if (false !== $type = get_resource_type($mixed)) {
+                    return new static($type);
+                }
+
+                throw new InvalidTransformationException('Could not transform resource to string.');
+            case 'null':
+                return new static('');
+            default:
+                throw new InvalidTransformationException('Could not transform unknown type to string');
+        }
     }
 }